Here it is. Have fun kids.
This is so not ready for general use yet it's not even funny. But if you have
a modicum of clue and the right need, you will probably find this immensely
useful.

What it does:

Let's you connect to your existing client with another IRC client, acting
as sort of a proxy. Currently only works with whatever network has your
focus currently (currently visible window, or the window with ^^^^^ in it.)
It will probably break if you connect and switch your input to another
network.

What works:
* Joining channels
* Sending PRIVMSG of all sort
* Sending CTCP
* Receiving PRIVMSG of all sort
* Receiving MODE changes
* Listing the channels you can join

What doesn't:
* Almost everything else
* Recieving CTCP responses
* Joining channels your client is not already in
* Multi-server
* Loading saved settings (not sure what's going on, maybe zak or kreca knows?)
* Whitelist (Anyone can connect, BE CAREFUL)

To Use:
	/addmod remote
	/config remote
	Set settings B, C and D
	/remote on
	Connect with another irc client

I've tested this with epic and some iphone client. If it doesn't work with
whatever crappy client you try it with let me know.